requirements.txt
streamlit
pandas
google-cloud-bigquery
matplotlib
altair
google-auth
db-dtypes
